requirements.txt
flask
flask-restplus
requests
numpy
matplotlib
keras
tensorflow==1.15.2
pillow
opencv-python
werkzeug==0.16.1
